+++ tools/Azure.Mcp.Tools.Extension/src/Constants.cs
@@ -6,4 +6,8 @@ namespace Azure.Mcp.Tools.Extension;
 internal class Constants
 {
     public const string AzureCliType = "az";
+
+    public const string AzureDeveloperCliType = "azd";
+
+    public const string AzureFunctionsCoreToolsCliType = "func";
 }